Move refreshAddon() into AddonReview
